requirements.txt
opencv-python==4.10.0.84
face-recognition==1.3.0
dlib==19.24.6
numpy==1.26.4
ultralytics==8.0.20
pyaudio==0.2.11
SpeechRecognition==3.10.0
scipy==1.13.0
mysql-connector-python==9.0.0
openpyxl==3.1.2
Pillow==10.4.0